requirements.txt
crewai
crewai_tools
load_dotenv
langchain-huggingface
langchain_google_genai
